# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/scheduling/settings": {
      "get": {
        "operationId": "SchedulingController_settings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarSettingsDto"
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [],
        "summary": "Consultar vista inicial del sistema",
        "tags": [
          "Calendario"
        ]
      },
      "patch": {
        "operationId": "SchedulingController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarSettingsDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Configurar vista inicial de client y backoffice",
        "tags": [
          "Calendario"
        ]
      }
    },
    "/api/scheduling/schedules": {
      "get": {
        "operationId": "SchedulingController_schedules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Listar horarios recurrentes",
        "tags": [
          "Calendario"
        ]
      },
      "post": {
        "description": "Crea o reemplaza la regla de la pareja. No elimina turnos ni solicitudes anteriores. Desactivar la regla oculta sus turnos automáticos.",
        "operationId": "SchedulingController_saveSchedule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Guardar horario recurrente por sede y deporte",
        "tags": [
          "Calendario"
        ]
      }
    },
    "/api/scheduling/blocked-days": {
      "get": {
        "operationId": "SchedulingController_blockedDays",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlockedDayResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Listar cierres de sedes",
        "tags": [
          "Calendario"
        ]
      },
      "post": {
        "description": "Afecta todos sus deportes, turnos manuales y automáticos. Impide seleccionar y generar WhatsApp; no envía cancelaciones ni elimina borradores.",
        "operationId": "SchedulingController_blockDay",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockedDayDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Operación completada. Respuesta directa, sin envoltorio data.",
            "headers": {
              "Cache-Control": {
                "description": "El borrador no debe cachearse.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlockedDayResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "409": {
            "description": "Duplicado o registro con dependencias que impiden la operación.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 409,
                  "message": "Duplicado o registro con dependencias que impiden la operación."
                }
              }
            }
          },
          "413": {
            "description": "El JSON supera 16 KiB.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 413,
                  "message": "El JSON supera 16 KiB."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Bloquear un día completo de una sede",
        "tags": [
          "Calendario"
        ]
      }
    },
    "/api/scheduling/blocked-days/{id}": {
      "delete": {
        "operationId": "SchedulingController_unblockDay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID público del registro.",
            "schema": {
              "format": "uuid",
              "example": "00000000-0000-4000-8000-000000000001",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Eliminado. Sin cuerpo de respuesta."
          },
          "400": {
            "description": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 400,
                  "message": "Datos inválidos, campos extra o nulos, selección incompatible o turno no disponible."
                }
              }
            }
          },
          "401": {
            "description": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 401,
                  "message": "X-API-Key ausente o incorrecta, o clave de gestión sin configurar."
                }
              }
            }
          },
          "403": {
            "description": "Origen no permitido por CORS.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 403,
                  "message": "Origen no permitido por CORS."
                }
              }
            }
          },
          "404": {
            "description": "Registro o relación referenciada inexistente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 404,
                  "message": "Registro o relación referenciada inexistente."
                }
              }
            }
          },
          "429": {
            "description": "Se superaron 100 solicitudes por minuto por IP.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 429,
                  "message": "Se superaron 100 solicitudes por minuto por IP."
                }
              }
            }
          },
          "503": {
            "description": "No se pudo acceder a la base de datos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                },
                "example": {
                  "statusCode": 503,
                  "message": "No se pudo acceder a la base de datos."
                }
              }
            }
          }
        },
        "security": [
          {
            "ManagementKey": []
          }
        ],
        "summary": "Reabrir un día bloqueado",
        "tags": [
          "Calendario"
        ]
      }
    },
...
  },
  "info": {
    "title": "API del turnero de polideportivos",
    "description": "Flujo: deporte → contacto → zona y sede → detalle → fecha y turno → vista previa de WhatsApp. Los catálogos empiezan vacíos. Cargarlos con clave de gestión en el orden deportes, zonas, sedes, relaciones sede/deporte y turnos. Los UUID de los ejemplos son ficticios. Generar WhatsApp no envía mensajes, no bloquea disponibilidad y no confirma reservas. Los borradores se acceden mediante UUID secreto, sin login. JSON de hasta 16 KiB; 100 solicitudes/minuto/IP.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "ManagementKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Clave de gestión solo para operaciones protegidas. No se almacena en la documentación."
      }
    },
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "Código de estado HTTP."
          },
          "message": {
            "description": "Mensaje de negocio o lista de errores de validación.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "El turno no está disponible"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Nombre del error; puede omitirse, por ejemplo en el límite de solicitudes."
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CalendarSettingsDto": {
        "type": "object",
        "properties": {
          "calendarEnabled": {
            "type": "boolean",
            "description": "Vista inicial calendario en client y backoffice."
          }
        },
        "required": [
          "calendarEnabled"
        ]
      },
      "ScheduleResponseDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "sportId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "weekdays": {
            "type": "number",
            "minimum": 1,
            "maximum": 127,
            "description": "Máscara semanal: domingo=1, lunes=2, ... sábado=64."
          },
          "opensAt": {
            "type": "string",
            "example": "09:00"
          },
          "closesAt": {
            "type": "string",
            "example": "22:00"
          },
          "durationMinutes": {
            "type": "number",
            "minimum": 15,
            "maximum": 720
          },
          "horizonDays": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "venueId",
          "sportId",
          "isActive",
          "weekdays",
          "opensAt",
          "closesAt",
          "durationMinutes",
          "horizonDays",
          "id"
        ]
      },
      "ScheduleDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "sportId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "weekdays": {
            "type": "number",
            "minimum": 1,
            "maximum": 127,
            "description": "Máscara semanal: domingo=1, lunes=2, ... sábado=64."
          },
          "opensAt": {
            "type": "string",
            "example": "09:00"
          },
          "closesAt": {
            "type": "string",
            "example": "22:00"
          },
          "durationMinutes": {
            "type": "number",
            "minimum": 15,
            "maximum": 720
          },
          "horizonDays": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          }
        },
        "required": [
          "venueId",
          "sportId",
          "isActive",
          "weekdays",
          "opensAt",
          "closesAt",
          "durationMinutes",
          "horizonDays"
        ]
      },
      "BlockedDayResponseDto": {
        "type": "object",
        "properties": {
          "venueId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "reason": {
            "type": "string",
            "maxLength": 120
          },
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "venueId",
          "date",
          "reason",
          "id"
        ]
      },
...
    }
  }
}